Move base classes to skbase.base
